+++ tests/test_database.py
@@ -132,7 +132,7 @@ class _BaseDatabaseTestCase:
         book = Book.last(name=name, author=author, pages_count=pages_count)
 
         assert isinstance(book, Book)
-        assert book._id == Book.count()
+        assert book.id
         assert book.name == name
         assert book.pages_count == pages_count