+++ docs/docs/release_notes.md
@@ -1,3 +1,11 @@
+### 4.0.0
+- Move `database` and `redis` connections from `MIDDLEWARES` to their own block, `DATABASE` and `REDIS`
+- Add `login()` & `logout()` to `JWTAuthentication` and used it in `BaseUser`
+- Support `Authentication` & `Authorization` in `Websocket`
+- Rename all exceptions suffix from `Exception` to `Error` (https://peps.python.org/pep-0008/#exception-names)
+- Support `pantherdb 1.4`
+- Minor Improvements
+
 ### 3.9.0
 - Change the style of `ModelSerializer` usage